% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{smv_get_remote_func}
\alias{smv_get_remote_func}
\title{Get a function for executing code on remote servers}
\usage{
smv_get_remote_func(server = NULL, user = NULL, ...)
}
\description{
This function takes in a remote server address and username, and
then it outputs a function that takes in an R-expression, executes
that expression on the remote server, and returns the output.
This is basicly a small wrapper around the remote_run function
}
