% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{smv_ggplot_discrete_cols}
\alias{smv_ggplot_discrete_cols}
\title{Get the colour values from the ggplot default discrete colour scale,
when using \code{n} distinct colours}
\usage{
smv_ggplot_discrete_cols(n)
}
\description{
Get the colour values from the ggplot default discrete colour scale,
when using \code{n} distinct colours
}
