% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tikz.R
\name{smv_latex_friendly_map_plot}
\alias{smv_latex_friendly_map_plot}
\title{Change the axes of a map plot so it becomes possible to plot using tikzDevice}
\usage{
smv_latex_friendly_map_plot(x)
}
\description{
Change the axes of a map plot so it becomes possible to plot using tikzDevice
}
