% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_png.R
\name{smv_pdf_convert}
\alias{smv_pdf_convert}
\title{Convert a pdf to png or jpeg}
\usage{
smv_pdf_convert(
  in_path,
  out_paths,
  format = c("png", "jpeg"),
  pages = NULL,
  dpi = 150,
  ...
)
}
\description{
Convert a pdf to png or jpeg
}
