% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{smv_progress_bar}
\alias{smv_progress_bar}
\title{Create a progress bar for tracking long-running processes.
This is a thin wrapper around the progress package}
\usage{
smv_progress_bar(n)
}
\description{
Create a progress bar for tracking long-running processes.
This is a thin wrapper around the progress package
}
