% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{smv_quiet}
\alias{smv_quiet}
\title{Turn an R expression completely quiet}
\usage{
smv_quiet(expr)
}
\description{
This function can be used for silencing the output of functions
that contain print statements inside, such as most of the
functions of the ncdf4 package.
}
