% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{smv_remote}
\alias{smv_remote}
\title{Execute code on a remote server}
\usage{
smv_remote(expr, user, server, substitute = TRUE, ...)
}
\description{
This function takes in an R-expression, a server name and a username for that server.
Then, it executes the R-expression on the remote server, using the future package.
Users are advised to setup passwordless login to the servers they are executing
their code on, to not have to retype their password every single time this function
is used.
}
