% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_command.R
\name{smv_run}
\alias{smv_run}
\title{Try to run command, and if it fails, then wait sleep_time seconds and try again}
\usage{
smv_run(command, args = character(), n_attempts = 1, sleep_time = 5, ...)
}
\description{
Try to run command, and if it fails, then wait sleep_time seconds and try again
}
